% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_compounds.R
\name{add_compound_name}
\alias{add_compound_name}
\title{Add Compound_Name and Annotation_Source columns to abundance data}
\usage{
add_compound_name(ab, annotation)
}
\description{
Add Compound_Name and Annotation_Source columns to abundance data
}
\keyword{internal}
