% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_compounds.R
\name{build_annotation_from_cache_and_user}
\alias{build_annotation_from_cache_and_user}
\title{Build annotation data.frame from cache, user library, and newly queried}
\usage{
build_annotation_from_cache_and_user(
  cas_vec,
  cache,
  user_lookup = list(),
  newly_queried = character(0)
)
}
\description{
Build annotation data.frame from cache, user library, and newly queried
}
\keyword{internal}
